requirements.txt
selenium
python-dotenv
schedule
pymongo
fuzz
rapidfuzz